+++ src/lib/zerodiv.mli
@@ -1,7 +1,0 @@
-(** ZeroDivision errors *)
-
-open IECCheckerCore
-module S = IECCheckerCore.Syntax
-
-val do_check : S.iec_library_element list -> Warn.t list
-val detector : Detector.t